Extract `PendingInteraction`, use `GLFW` mouse constants
